refactor: BindingAdapter 메서드와 Mapper 메서드 분리
